HW beginners packages' structure was reorganized
