+++ src/javaBeginners/HW6.java
@@ -4,69 +4,13 @@ import Common.Printing;
 
 public class HW6 {
 
-    public static void main(String[] args) {
-
-        Printing.printTask("1");
-        System.out.println("Print sequence [0, 9]");
-        descSeq();
-        System.out.println();
-
-        Printing.printTask("2");
-        System.out.println("Print subsequence (10, 0]");
-        ascSeq();
-        System.out.println();
-
-        Printing.printTask("3");
-        System.out.println("Print seq [50, 55] step width 2");
-        seq3();
-        System.out.println();
-
-        Printing.printTask("4");
-        System.out.println("Print seq (327, 300) multiple of 7");
-        seq4();
-        System.out.println();
-
-        Printing.printTask("5");
-        System.out.println("Print seq [12,13] step width 0.1");
-        seq5();
-        System.out.println();
-
-        Printing.printTask("6");
-        System.out.println("Print seq [215, 237] multiple of 2");
-        seq6();
-        System.out.println();
-
-        Printing.printTask("7");
-        System.out.println("Print seq (7, 14) multiple of 7");
-        seq7();
-        System.out.println();
-
-        Printing.printTask("8");
-        System.out.println("Print seq [-32,768, 32,768] multiple of 15001");
-        seq8();
-        System.out.println();
-
-        Printing.printTask("9");
-        System.out.println("Print seq [-10, 34] multiple of 11 print blue, multiple of 12 print red, 0 is ZERO");
-        seq9();
-        System.out.println();
-
-        Printing.printTask("10");
-        System.out.println("Print user seq with input params start, end, step");
-        System.out.println(userSeq(16, 75, 4));
-        System.out.println();
-
-    }
-
     public static void descSeq(){
-
         for ( int i = 0; i <= 9; i++){
             System.out.print(i);
         }
     }
 
     public static void ascSeq() {
-
         for (int i = 9; i > 0; i--) {
             System.out.print(i);
         }
@@ -80,7 +24,6 @@ public class HW6 {
     }
 
     public static void seq4(){
-
         for (int i = 326; i > 300; i--) // (327, 300)
             if (i % 7 == 0) {
                 System.out.print(i + " ");
@@ -89,7 +32,6 @@ public class HW6 {
     }
 
     public static void seq5(){
-
         double n = 0.1;
         for ( double i = 13; i >= 12; i -= n ){
             System.out.print(String.format("%,.1f", i) + " ");
@@ -98,7 +40,6 @@ public class HW6 {
     }
 
     public static void seq6(){
-
         for (int i = 215; i <= 237; i++){
             if (i % 2 == 0){
                 System.out.print(i + " ");
@@ -107,7 +48,6 @@ public class HW6 {
     }
 
     public static void seq7(){
-
         for (int i = 8; i <= 14; i++)
             if(i % 7 ==0){ // always false
                 System.out.print(i +" ");
@@ -116,9 +56,7 @@ public class HW6 {
     }
 
     public static void seq8(){
-
         for (int i = Short.MIN_VALUE; i < Short.MAX_VALUE; i++ ){
-
             if ( i % 15001 == 0){
                 System.out.print(i + " ");
             }
@@ -127,7 +65,6 @@ public class HW6 {
     }
 
     public static void seq9(){
-
         for (int i = -10; i <= 34; i++){
             if (i == 0){
                 System.out.print(" zero ");
@@ -136,32 +73,21 @@ public class HW6 {
                 if (i % 12 == 0){
                     System.out.print("\u001B[31m" + i +" \u001B[0m");
                 }
-
-
                     if (i % 11 == 0) {
                         System.out.print("\u001B[34m" + i + " \u001B[0m");
                     }
-
                 else System.out.print(i + " ");
             }
-
         }
     }
 
     public static String userSeq(int start, int end, int step){ // task 10
         String s = "";
-
         for (int i = start; i <= end; i +=step){
-            s = s + i + " ";
+            s += i + " ";
         }
         return s;
-
-
     }
-
-
-
-
 
 }
