+++ src/javaBeginners/HW8.java
@@ -1,0 +1,4 @@
+package javaBeginners;
+
+public class HW8 {
+}
